Adding helper for fetching DALI examples if provided by a TAP service
Adding support for /examples as defined in http://www.ivoa.net/documents/DALI/index.html 1.1. This is not required by the TAP standard, but if a service wants to provide its own example queries, this is the format to use. These qieroes can be particularly helpful for validation testing, and if user-defined-functions are permitted or geometry queries are supported on some but not all tables.